requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
langchain==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langgraph==0.2.60
